[RC-9] Moved capitalizer decorator
